+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Current
+# 0.5.3
 
 - Fix download responses being buffered in CLI
 - Test for `get_trading_pair_page_url()` in the downloaded dataset